+++ CHANGELOG.md
@@ -31,7 +31,9 @@ Scientific Changes section), and this project adheres to [Semantic Versioning](h
   The refactoring also required some small user facing changes:
   - The `regions_of_interest` parameter of `RoiStrideSegmentation.segment` is now keyword only to be compatible with the
     `StrideSegmentation` base class.
-  
+  - The `UllrichGaitSequenceDetection` now returns a dict also in the case of merged gait sequences. The gait
+   sequences stored for all sensor are set equal. 
+   https://mad-srv.informatik.uni-erlangen.de/MadLab/GaitAnalysis/gaitmap/-/merge_requests/121
 
 ### Deprecated